+++ shopify/utils/tsup.config.ts
@@ -1,6 +1,6 @@
 import path from "node:path";
 import fs from "node:fs/promises";
-// import jsDelivrImportsPlugin from "@repo/js-delivr-imports-plugin";
+import { jsDelivrImportsPlugin } from "@repo/js-delivr-imports-plugin";
 import { defineConfig } from "tsup";
 import { generateDtsBundle } from "dts-bundle-generator";
 
@@ -32,7 +32,7 @@ export default defineConfig([
 
       await fs.writeFile(
         path.resolve(outDir, "effect", "index.d.ts"),
-        effectDts
+        effectDts,
       );
     },
   },
@@ -48,7 +48,7 @@ export default defineConfig([
             filePath: path.resolve("./src/index.ts"),
           },
         ],
-        { preferredConfigPath: "./tsconfig.json" }
+        { preferredConfigPath: "./tsconfig.json" },
       );
 
       if (!dts) return;
@@ -56,27 +56,13 @@ export default defineConfig([
       await fs.writeFile(path.resolve(outDir, "main", "index.d.ts"), dts);
     },
   },
-  // {
-  //   ...commonConfig,
-  //   format: "esm",
-  //   entry: ["src/**/*.ts"],
-  //   outDir: `${outDir}/jsdelivr`,
-  //   dts: false,
-  //   async onSuccess() {
-  //     const [indexDts] = generateDtsBundle([
-  //       {
-  //         filePath: path.resolve("./src/index.ts"),
-  //       },
-  //     ]);
+  {
+    ...commonConfig,
+    format: "esm",
+    entry: ["src/**/*.ts"],
+    outDir: `${outDir}/jsdelivr`,
+    dts: false,
 
-  //     if (!indexDts) return;
-
-  //     await fs.writeFile(
-  //       path.resolve(outDir, "jsdelivr", "index.d.ts"),
-  //       indexDts
-  //     );
-  //   },
-
-  //   esbuildPlugins: [jsDelivrImportsPlugin()],
-  // },
+    esbuildPlugins: [jsDelivrImportsPlugin()],
+  },
 ]);