+++ packages/web-components/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @brytdesigns/web-component-utils
 
+## 0.1.3
+
+### Patch Changes
+
+- Removed typegen for jsdelivr and fixed export paths as well
+
 ## 0.1.2
 
 ### Patch Changes