+++ shopify/packages/hybrid-cart/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @brytdesigns/shopify-hybrid-cart
 
+## 0.1.3
+
+### Patch Changes
+
+- Removed typegen for jsdelivr and fixed export paths as well
+- Updated dependencies
+  - @brytdesigns/shopify-hybrid-cart-api@1.2.3
+
 ## 0.1.2
 
 ### Patch Changes